+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -5,10 +5,10 @@ package info.metadude.kotlin.library.engelsystem
 object GradlePlugins {
 
     private object Versions {
-        const val dokka = "1.9.20"
-        const val kotlin = "2.0.21"
-        const val ksp = "2.0.21-1.0.27"
-        const val versions = "0.51.0"
+        const val dokka = "2.0.0"
+        const val kotlin = "2.1.20"
+        const val ksp = "2.1.20-2.0.1"
+        const val versions = "0.52.0"
     }
 
     const val dokka = "org.jetbrains.dokka:dokka-gradle-plugin:${Versions.dokka}"
@@ -20,17 +20,19 @@ object GradlePlugins {
 object Libs {
 
     private object Versions {
-        const val junitJupiter = "5.11.3"
-        const val kotlinCoroutines = "1.9.0"
-        const val moshi = "1.15.1"
+        const val junitJupiter = "5.12.2"
+        const val junitPlatformLauncher = "1.12.2"
+        const val kotlinCoroutines = "1.10.2"
+        const val moshi = "1.15.2"
         const val okhttp = "4.12.0"
         const val retrofit = "2.11.0"
-        const val threetenbp = "1.7.0"
+        const val threetenbp = "1.7.1"
         const val truth = "1.4.4"
     }
 
     const val junitJupiterApi = "org.junit.jupiter:junit-jupiter-api:${Versions.junitJupiter}"
     const val junitJupiterEngine = "org.junit.jupiter:junit-jupiter-engine:${Versions.junitJupiter}"
+    const val junitPlatformLauncher = "org.junit.platform:junit-platform-launcher:${Versions.junitPlatformLauncher}"
     const val kotlinCoroutinesTest = "org.jetbrains.kotlinx:kotlinx-coroutines-test:${Versions.kotlinCoroutines}"
     const val moshi = "com.squareup.moshi:moshi:${Versions.moshi}"
     const val moshiKotlinCodegen = "com.squareup.moshi:moshi-kotlin-codegen:${Versions.moshi}"